requirements.txt
alembic==1.8.1
aniso8601==9.0.1
astroid==2.12.13
Authlib==1.2.0
click==8.1.3
cryptography==38.0.4
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Migrate==2.6.0
Flask-Moment==1.0.5
Flask-RESTful==0.3.9
Flask-Script==2.0.5
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
future==0.18.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.8.0
Mako==1.2.4
MarkupSafe==2.1.1
mccabe==0.7.0
postgres==4.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
psycopg2-pool==1.1
pyasn1==0.4.8
PyJWT==1.7.1
pylint==2.15.8
python-dateutil==2.8.2
python-dotenv==0.21.0
python-editor==1.0.4
python-jose==3.3.0
python-jose-cryptodome==1.3.2
rsa==4.9
typed-ast==1.5.4
six==1.16.0
SQLAlchemy==1.4.44
sqlalchemy-orm==1.2.3
urllib3==1.26.13
Werkzeug==2.2.2
wrapt==1.14.1